Create This is my Third Task by Aderonke
For this assignment, I had to create an array that stores traffic light colors and then use control flow to decide what action a car should take depending on the light.

I started by defining an array with three colors: Green, Yellow, and Red. After that, I used an if-else statement to check the current color. If the light is Green, the car should move. If it’s Amber, the car should get ready to go. And if it’s Red, the car should stop.

To make the code cleaner, I also assigned the current light to a variable instead of checking the array directly each time. This makes it easier to read and manage.

In the end, the program clearly prints out what the car should do based on the color of the traffic light.